fix some issue and add docker-compose dockerfile

+++ src/main/java/com/twt/club/registration/entity/Registration.java
@@ -28,5 +28,5 @@ public class Registration {
     private LocalDateTime updatedAt;
 
     @TableLogic
-    private long isDeleted;
+    private Integer isDeleted;
 }
